+++ test/multinetwork.jl
@@ -207,7 +207,7 @@ TESTLOG = Memento.getlogger(PowerModels)
             result = PowerModels.run_mn_opf(mn_data, SparseSDPWRMPowerModel, sdp_solver)
 
             # hits iteration limit in SCS v0.9, sense ALMOST_OPTIMAL
-            @test result["termination_status"] == ALMOST_OPTIMAL
+            @test result["termination_status"] == OPTIMAL || result["termination_status"] == ALMOST_OPTIMAL
             # tolerance relaxed for cross platform compat.
             @test isapprox(result["objective"], 33321.9; atol = 1e2)
             @test isapprox(