+++ CHANGELOG.md
@@ -3,6 +3,7 @@ PowerModels.jl Change Log
 
 ### Staged
 - Add specialized version of `sol_component_fixed` for PowerModels (#821)
+- Add support for Memento v1.4
 - Fix `tnep` ne_branch solution values to be consistent with regular branches (#820)
 
 ### v0.19.5